refactor: trim large command modules into cohesive siblings (#322)
Split the three largest src modules so each command __init__ stays focused
on orchestration, mirroring the existing bump/ and release/ layout. No
behaviour change — extracted helpers are re-exported so the public import
surface (and the monkeypatch paths tests rely on) is unchanged.

- rollback/__init__.py 421 -> 269: execution engine moved to
  rollback/engine.py, RollbackOptions to rollback/models.py, the summary
  printer to rollback/io.py.
- bump/__init__.py 407 -> 347: _resolve_language moved to bump/models.py,
  _resolve_bump_baseline to bump/io.py (conftest + #1126 guard patches
  repointed to where get_latest_remote_version is now looked up).
- cli.py 385 -> 334: dropped the docstring Args: blocks that duplicated the
  typer.Option(help=...) text verbatim.

All src modules are now <=347 lines; radon avg CC stays A (3.49) with no
C-or-worse blocks. Gates green: fmt, typecheck (ty + mypy --strict),
docs-coverage 100%, deptry, security, validate, test (493 passed, 100% cov).
