+++ src/rhiza_tools/cli.py
@@ -92,18 +92,8 @@ def bump(
     projects using semantic versioning. You can provide an explicit version number,
     a bump type (patch, minor, major), or leave it blank for an interactive prompt.
 
-    Args:
-        version: The version to bump to. Can be an explicit version (e.g., "1.2.3"),
-            a bump type ("patch", "minor", "major"), a prerelease type
-            ("alpha", "beta", "rc", "dev"), or None for interactive selection.
-        language: Programming language (python or go). Auto-detected if not specified.
-        dry_run: If True, show what would change without actually changing anything.
-        commit: If True, automatically commit the version change to git.
-        push: If True, push changes to remote after commit (implies --commit).
-        branch: Branch to perform the bump on (default: current branch).
-        allow_dirty: If True, allow bumping even with uncommitted changes.
-        config: Path to the .cfg.toml bumpversion config file. Defaults to .rhiza/.cfg.toml.
-        verbose: If True, enable verbose debug output.
+    Each argument is documented by its ``--help`` text above; see
+    :func:`rhiza_tools.commands.bump.bump_command` for the underlying behaviour.
     """
     _apply_verbose(verbose)
     lang_enum = parse_language_option(language)
@@ -142,12 +132,7 @@ def generate_coverage_badge(
 
     Reads a coverage report JSON file and creates a shields.io endpoint JSON file
     for displaying a coverage badge. The badge color automatically adjusts based
-    on the coverage percentage.
-
-    Args:
-        coverage_json: Path to the coverage.json file generated by pytest-cov.
-        output: Path where the badge JSON file should be written.
-        verbose: If True, enable verbose debug output.
+    on the coverage percentage. Each argument is documented by its ``--help`` text.
     """
     _apply_verbose(verbose)
     generate_coverage_badge_command(coverage_json_path=coverage_json, output_path=output)
@@ -181,16 +166,8 @@ def release(
     Supports Python projects (pyproject.toml) and Go projects (go.mod + VERSION
     file). The project language is auto-detected when not explicitly specified.
 
-    Args:
-        bump: Bump type (MAJOR, MINOR, PATCH) to apply. Selected interactively when omitted.
-        push: If True, push changes without prompting (implies non-interactive for push).
-        dry_run: If True, show what would happen without actually pushing the tag.
-        non_interactive: If True, skip all confirmation prompts and default the bump to
-            patch when no --bump type is given (useful for CI/CD).
-        language: Programming language (python or go). Auto-detected if not specified.
-        allow_older: If True, allow releasing a version not newer than the latest remote release.
-        config: Path to the .cfg.toml bumpversion config file. Passed through to the bump.
-        verbose: If True, enable verbose debug output.
+    Each argument is documented by its ``--help`` text above; see
+    :func:`rhiza_tools.commands.release.release_command` for the underlying behaviour.
     """
     _apply_verbose(verbose)
     lang_enum = parse_language_option(language)
@@ -215,14 +192,8 @@ def rollback(
     It uses ``git revert`` rather than ``git reset``, making it safe
     even when changes have already been pushed to remote.
 
-    Args:
-        tag: The tag to rollback (e.g., "v1.2.3"). If omitted, an interactive
-            menu shows recent tags to choose from.
-        revert_bump: If True, also revert the version bump commit associated
-            with the tag.
-        dry_run: If True, show what would happen without actually making changes.
-        non_interactive: If True, skip all confirmation prompts (useful for CI/CD).
-        verbose: If True, enable verbose debug output.
+    Each argument is documented by its ``--help`` text above; see
+    :func:`rhiza_tools.commands.rollback.rollback_command` for the underlying behaviour.
     """
     _apply_verbose(verbose)
     options = RollbackOptions(
@@ -243,11 +214,7 @@ def update_readme(
 
     This command runs `make help` and updates the README.md file with the current
     help output, keeping the documentation in sync with available Make targets.
-
-    Args:
-        dry_run: If True, show the help output that would be inserted without
-            actually modifying README.md.
-        verbose: If True, enable verbose debug output.
+    Each argument is documented by its ``--help`` text above.
     """
     _apply_verbose(verbose)
     update_readme_command(dry_run)
@@ -275,13 +242,8 @@ def version_matrix(
 
     This command reads the requires-python field from pyproject.toml and outputs
     a JSON array of Python versions that satisfy the constraint. This is primarily
-    used in GitHub Actions to compute the test matrix.
-
-    Args:
-        pyproject: Path to the pyproject.toml file.
-        candidates: Comma-separated list of candidate Python versions to evaluate.
-            Defaults to "3.11,3.12,3.13,3.14".
-        verbose: If True, enable verbose debug output.
+    used in GitHub Actions to compute the test matrix. Each argument is documented
+    by its ``--help`` text above.
     """
     _apply_verbose(verbose)
     candidates_list = None
@@ -325,11 +287,7 @@ def analyze_benchmarks(
     Note: This command requires pandas and plotly. Install with:
     uv pip install -e '.[dev]' or pip install 'rhiza-tools[dev]'
 
-    Args:
-        benchmarks_json: Path to the benchmarks.json file.
-        output_html: Path where the HTML visualization should be saved.
-        show: If True, open the interactive chart in a browser after saving.
-        verbose: If True, enable verbose debug output.
+    Each argument is documented by its ``--help`` text above.
     """
     _apply_verbose(verbose)
     analyze_benchmarks_command(benchmarks_json=benchmarks_json, output_html=output_html, show=show)
@@ -346,10 +304,6 @@ def pip_audit(ctx: typer.Context, verbose: bool = VERBOSE_OPTION) -> None:
     tooling (pip, setuptools, wheel, distribute) warn without failing. Any extra
     arguments after the command are forwarded verbatim to pip-audit
     (e.g. ``rhiza-tools pip-audit --ignore-vuln CVE-2024-1234``).
-
-    Args:
-        ctx: Typer context; its extra args are forwarded verbatim to pip-audit.
-        verbose: If True, enable verbose debug output.
     """
     _apply_verbose(verbose)
     raise typer.Exit(code=pip_audit_command(ctx.args))
@@ -375,11 +329,6 @@ def suppression_audit(
     histogram, and a letter grade. With ``--fail-stale-nosec-cve`` it also
     cross-checks CVE-tagged # nosec comments against live pip-audit findings and
     fails on stale ones; extra arguments are forwarded to pip-audit.
-
-    Args:
-        ctx: Typer context; its extra args are forwarded to pip-audit.
-        fail_stale_nosec_cve: If True, fail on stale CVE-tagged # nosec comments.
-        verbose: If True, enable verbose debug output.
     """
     _apply_verbose(verbose)
     raise typer.Exit(code=suppression_audit_command(fail_stale_nosec_cve, ctx.args))
